test(agent): isolate response constructor regression
